+++ CHANGELOG.md
@@ -7,6 +7,8 @@ New features:
 Changes:
 - Rename `File` capture device to `RawFile`.
 - Filter pipeline steps take a list of channels to filter instead of a single one.
+Bugfixes:
+- MacOS: Support devices appearing as separate capture and playback devices.
 
 ## v2.0.3
 Bugfixes: